feat(automations): surface the substitutions "?" on the builder, not just the Test panel (#3653)
The "all substitutions" help was only reachable from inside the ▶ Test panel.
Move the drawer + token tables into a shared SubstitutionsHelp component and add
a "?" affordance (with a one-line tip) at the top of the AutomationBuilder, where
authors actually write {{ trigger.* }} / {{ var.* }} in message/notify fields.
The Test panel reuses the same shared drawer.
