+++ src/components/automations/AutomationTester.tsx
@@ -10,6 +10,7 @@
 import { useState, type ReactNode } from 'react';
 import apiService from '../../services/api';
 import type { VariableOption } from './AutomationBuilder';
+import SubstitutionsHelpDrawer from './SubstitutionsHelp';
 
 export interface SimResult {
   matched: boolean;
@@ -257,65 +258,6 @@ function ActionView({ a }: { a: SimResult['actions'][number] }) {
             </details>
           </>}
     </div>
-  );
-}
-
-// All `{{ trigger.* }}` tokens, by trigger type. `sourceId`/`timestamp` are added to every group.
-const TRIGGER_TOKENS: Record<string, Array<[string, string]>> = {
-  'trigger.message': [
-    ['text', 'Message body'], ['from', 'Sender node number'], ['fromId', 'Sender node id (!hex)'],
-    ['to', 'Recipient node number'], ['toId', 'Recipient node id'], ['channel', 'Channel index'],
-    ['portnum', 'Port number'], ['packetId', 'Packet id (used as tapback replyId)'],
-    ['hops', 'Hop count (hopStart − hopLimit)'], ['hopStart', 'Hop start'], ['hopLimit', 'Hop limit'],
-    ['snr', 'Receive SNR — RF-received messages only'], ['rssi', 'Receive RSSI dBm — RF only'],
-    ['isDM', 'true if a direct message'], ['isBroadcast', 'true if broadcast'],
-    ['wantAck', 'Sender requested an ack'], ['replyId', 'Replied-to packet id'],
-    ['emoji', 'Tapback/reaction emoji flag'], ['viaMqtt', 'true if it arrived via MQTT'],
-    ['decryptedBy', 'Channel/key that decrypted it'],
-  ],
-  'trigger.telemetry': [['nodeNum', 'Node number'], ['telemetryType', 'Metric name'], ['value', 'Reading value'], ['unit', 'Unit']],
-  'trigger.nodeUpdated': [['nodeNum', 'Node number'], ['changed', 'Changed field names (list)']],
-  'trigger.nodeDiscovered': [['nodeNum', 'Node number'], ['changed', 'Changed field names (list)']],
-  'trigger.system': [['event', 'System event'], ['nodeNum', 'Node number (if any)'], ['reason', 'Detail / reason'], ['latestVersion', 'Latest version (upgrade-available)'], ['currentVersion', 'Current version (upgrade-available)']],
-  'trigger.geofence': [['event', 'enter / exit / dwell'], ['nodeNum', 'Node number'], ['latitude', 'Node latitude'], ['longitude', 'Node longitude'], ['distanceKm', 'Distance from the region centre (km)']],
-  'trigger.schedule': [],
-};
-const UNIVERSAL_TOKENS: Array<[string, string]> = [['sourceId', 'The source the event came from'], ['timestamp', 'Event time (epoch ms)']];
-const TRIGGER_LABEL: Record<string, string> = {
-  'trigger.message': 'Message', 'trigger.telemetry': 'Telemetry', 'trigger.nodeUpdated': 'Node updated',
-  'trigger.nodeDiscovered': 'Node discovered', 'trigger.system': 'System event', 'trigger.geofence': 'Geofence', 'trigger.schedule': 'Schedule',
-};
-
-/** Drawer listing every available substitution token. */
-function SubstitutionsHelpDrawer({ triggerType, variables, onClose }: { triggerType: string; variables: VariableOption[]; onClose: () => void }) {
-  const order = [triggerType, ...Object.keys(TRIGGER_TOKENS).filter((t) => t !== triggerType)];
-  return (
-    <>
-      <div className="ae-drawer-overlay" onClick={onClose} />
-      <div className="ae-drawer">
-        <button className="ae-btn ae-btn--ghost ae-drawer-close" onClick={onClose}>✕</button>
-        <h2>Substitutions</h2>
-        <p className="ae-muted">Insert these <code>{'{{ … }}'}</code> tokens in any text field (message, notify title/body). An unknown or empty value renders blank.</p>
-
-        <h3>Variables &amp; misc</h3>
-        <dl>
-          <dt>{'{{ var.NAME }}'}</dt><dd>Any user variable{variables.length ? `: ${variables.map((v) => v.name).join(', ')}` : ' (none defined yet)'}.</dd>
-          <dt>{'{{ NOW }}'}</dt><dd>Current time (epoch ms).</dd>
-        </dl>
-
-        {order.filter((t) => TRIGGER_TOKENS[t]).map((t) => (
-          <div key={t}>
-            <h3>{TRIGGER_LABEL[t] ?? t}{t === triggerType ? ' — current trigger' : ''}</h3>
-            <dl>
-              {[...TRIGGER_TOKENS[t], ...UNIVERSAL_TOKENS].flatMap(([k, d]) => [
-                <dt key={`${k}-t`}>{`{{ trigger.${k} }}`}</dt>,
-                <dd key={`${k}-d`}>{d}</dd>,
-              ])}
-            </dl>
-          </div>
-        ))}
-      </div>
-    </>
   );
 }
